additional cleanup on xstreamcdn extractor
